check permission role & add new api for App

+++ ruoyi-admin/src/main/java/com/ruoyi/web/controller/fx67ll/lottery/Fx67llLotteryLogController.java
@@ -3,6 +3,7 @@ package com.ruoyi.web.controller.fx67ll.lottery;
 import java.util.List;
 import javax.servlet.http.HttpServletResponse;
 
+import com.ruoyi.common.utils.SecurityUtils;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -80,6 +81,20 @@ public class Fx67llLotteryLogController extends BaseController {
     }
 
     /**
+     * 提供给 APP 获取每日号码记录详细信息
+     */
+//    如果只放开SecurityConfig中允许匿名请求的配置，不放开这里的权限配置，会返回获取用户信息异常的错误
+//    @PreAuthorize("@ss.hasPermi('lottery:log:query')")
+    @GetMapping(value = "/getLotteryLogInfoForApp/{lotteryId}")
+    public AjaxResult getLotteryLogInfoForApp(@PathVariable("lotteryId") Long lotteryId) {
+        if (fx67llLotteryLogService.selectFx67llLotteryLogByLotteryId(lotteryId).getUserId() != SecurityUtils.getUserId()) {
+            return warn("此记录非本人创建，禁止查询！");
+        } else {
+            return success(fx67llLotteryLogService.selectFx67llLotteryLogByLotteryId(lotteryId));
+        }
+    }
+
+    /**
      * 新增每日号码记录
      */
     @PreAuthorize("@ss.hasPermi('lottery:log:add')")
@@ -111,6 +126,21 @@ public class Fx67llLotteryLogController extends BaseController {
     }
 
     /**
+     * 提供给 APP 修改每日号码记录
+     */
+//    如果只放开SecurityConfig中允许匿名请求的配置，不放开这里的权限配置，会返回获取用户信息异常的错误
+//    @PreAuthorize("@ss.hasPermi('lottery:log:edit')")
+    @Log(title = "每日号码记录", businessType = BusinessType.UPDATE)
+    @PostMapping("/editLotteryLogForApp")
+    public AjaxResult editLotteryLogForApp(@RequestBody Fx67llLotteryLog fx67llLotteryLog) {
+        if (fx67llLotteryLogService.selectFx67llLotteryLogByLotteryId(fx67llLotteryLog.getLotteryId()).getUserId() != SecurityUtils.getUserId()) {
+            return warn("此记录非本人创建，禁止修改！");
+        } else {
+            return toAjax(fx67llLotteryLogService.updateFx67llLotteryLog(fx67llLotteryLog));
+        }
+    }
+
+    /**
      * 删除每日号码记录
      */
     @PreAuthorize("@ss.hasPermi('lottery:log:remove')")
@@ -128,7 +158,11 @@ public class Fx67llLotteryLogController extends BaseController {
     @Log(title = "每日号码记录", businessType = BusinessType.DELETE)
     @DeleteMapping("/deleteLogByIdForApp/{lotteryId}")
     public AjaxResult deleteLogByIdForApp(@PathVariable Long lotteryId) {
-        return toAjax(fx67llLotteryLogService.deleteFx67llLotteryLogByLotteryIdForApp(lotteryId));
+        if (fx67llLotteryLogService.selectFx67llLotteryLogByLotteryId(lotteryId).getUserId() != SecurityUtils.getUserId()) {
+            return warn("此记录非本人创建，禁止删除！");
+        } else {
+            return toAjax(fx67llLotteryLogService.deleteFx67llLotteryLogByLotteryIdForApp(lotteryId));
+        }
     }
 
 }
